requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
requests>=2.31.0
plotly>=5.15.0
reportlab>=4.0.4
markdown>=3.4.4
streamlit>=1.28.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
streamlit
jinja2
plotly

